Explicitly handle daily_stats average times as floats

+++ src/chia_log/handlers/daily_stats/stat_accumulators/search_time_stats.py
@@ -9,20 +9,20 @@ class SearchTimeStats(HarvesterActivityConsumer, StatAccumulator):
     def __init__(self):
         self._last_reset_time = datetime.now()
         self._num_measurements = 0
-        self._avg_time_seconds = 0
+        self._avg_time_seconds = 0.0
         self._over_5_seconds = 0
         self._over_15_seconds = 0
 
     def reset(self):
         self._last_reset_time = datetime.now()
         self._num_measurements = 0
-        self._avg_time_seconds = 0
+        self._avg_time_seconds = 0.0
         self._over_5_seconds = 0
         self._over_15_seconds = 0
 
     def consume(self, obj: HarvesterActivityMessage):
         self._num_measurements += 1
-        self._avg_time_seconds += int((obj.search_time_seconds - self._avg_time_seconds) / self._num_measurements)
+        self._avg_time_seconds += (obj.search_time_seconds - self._avg_time_seconds) / self._num_measurements
         if obj.search_time_seconds > 5:
             self._over_5_seconds += 1
         if obj.search_time_seconds > 15:
